04-Weather-Agent-weather-ApiTool/pyproject.toml
[project]
name = "04-agent-tools-callig"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.0.13",
    "python-dotenv>=1.1.0",
]
